+++ packages/rum-recorder/src/boot/recorder.ts
@@ -29,9 +29,9 @@ export function startRecording(
     emit: addRawRecord,
   })!
 
-  lifeCycle.subscribe(LifeCycleEventType.SESSION_RENEWED, takeFullSnapshot)
   lifeCycle.subscribe(LifeCycleEventType.VIEW_CREATED, takeFullSnapshot)
   const { stop: stopTrackingFocusRecords } = trackFocusRecords(lifeCycle, addRawRecord)
+  trackViewEndRecord(lifeCycle, (record) => addRawRecord(record))
 
   return {
     stop() {
@@ -54,4 +54,12 @@ export function trackFocusRecords(lifeCycle: LifeCycle, addRawRecord: (record: R
   addFocusRecord()
   lifeCycle.subscribe(LifeCycleEventType.VIEW_CREATED, addFocusRecord)
   return addEventListeners(window, [DOM_EVENT.FOCUS, DOM_EVENT.BLUR], addFocusRecord)
+}
+
+export function trackViewEndRecord(lifeCycle: LifeCycle, addRawRecord: (record: RawRecord) => void) {
+  lifeCycle.subscribe(LifeCycleEventType.VIEW_ENDED, () => {
+    addRawRecord({
+      type: RecordType.ViewEnd,
+    })
+  })
 }