+++ packages/rum-recorder/src/domain/rrweb/observer.ts
@@ -1,5 +1,5 @@
 import { noop } from '@datadog/browser-core'
-import { INode, MaskInputOptions, SlimDOMOptions } from 'rrweb-snapshot'
+import { INode, MaskInputOptions, SlimDOMOptions } from '../rrweb-snapshot'
 import { MutationBuffer } from './mutation'
 import {
   Arguments,
@@ -28,6 +28,7 @@ import {
   ViewportResizeCallback,
 } from './types'
 import {
+  forEach,
   getWindowHeight,
   getWindowWidth,
   hookSetter,
@@ -206,7 +207,7 @@ function initInputObserver(
     // the other radios with the same name attribute will be unchecked.
     const name: string | undefined = (target as HTMLInputElement).name
     if (type === 'radio' && name && isChecked) {
-      document.querySelectorAll(`input[type="radio"][name="${name}"]`).forEach((el) => {
+      forEach(document.querySelectorAll(`input[type="radio"][name="${name}"]`), (el: Element) => {
         if (el !== target) {
           cbWithDedup(el, {
             isChecked: !isChecked,